# ProVR-APIs-1.2.0-swagger.yaml
openapi: 3.0.0
info:
  title: Microservice API's
  description: The APIs of the ProVR virtual reality application
  license:
    name: Open Source (MIT Licence)
  version: 1.2.0
servers:
- url: https://virtserver.swaggerhub.com/provr/ProVR-APIs/1.2.0
  description: SwaggerHub API Auto Mocking
- url: https://{deploymentHost}
  description: Production server
  variables:
    deploymentHost:
      description: Host to be used. This is the URL of the API gateway.
      default: https://api-gateway-iu3tuzfidq-ez.a.run.app
...
paths:
...
  /auth/auth_check:
    post:
      tags:
      - auth
      summary: Authorization check
      description: Checks if the provided basic token in the autorization header is valid or not. Only if this succeeds the API gateway will forward the client request to the target endpoint.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth_check_body'
      responses:
        "401":
          description: Unauthorized. Please re-login
        "405":
          description: Only POST method is accepted
        "500":
          description: Internal error
        "200":
          description: The access check was successful. The token is authorized for the path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/simple_success_response_200'
  /auth/{username}:
    delete:
      tags:
      - auth
      summary: Delete an account
      description: Removes a specific account associated with ProVR. This endpoint is currently not supported in production (needs verification that the user performing the deletion is the owner).
      parameters:
      - name: username
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "401":
          description: Unauthorized. Please re-login
        "404":
          description: User does not exist
        "500":
          description: Internal error
        "200":
          description: The deletion was successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/simple_success_response_200'
...
components:
  schemas:
...
    auth_check_body:
      required:
      - token
      type: object
      properties:
        token:
          type: string
          description: Basic token needed for authorization of resource access. Expires after 30 minutes
    simple_success_response_200:
      type: object
      properties:
        message:
          type: string
          description: Success message
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic